+++ index.html
@@ -2,7 +2,7 @@
 <html lang="en">
 	<head>
 		<meta charset="UTF-8" />
-		<!--<link rel="icon" type="image/svg+xml" href="/vite.svg" />-->
+		<link rel="icon" type="image/svg+xml" href="/src/assets/react.svg" />
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
 		<title>Bloxes</title>
 	</head>
