+++ src/components/contextMenu/PlainContext.tsx
@@ -2,9 +2,13 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { useSelector, useDispatch } from 'react-redux';
 import FolderContext from './FolderContext';
-import FileContext from './FileContext';
+import FileContext from '@/components/contextMenu/FileContext';
 import MainContext from './MainContext.js';
+import { store } from '@/redux/store';
 import { setRenderConditionFalse } from '../../redux/contextSlice';
+import ContextViewModel from '@/components/contextMenu/ContextViewModel';
+
+const logic = new ContextViewModel(store);
 
 PlainContext.propTypes = {
 	top: PropTypes.number,
@@ -12,8 +16,6 @@ PlainContext.propTypes = {
 };
 
 export default function PlainContext(props) {
-	const dispatch = useDispatch();
-
 	const selectedFile = useSelector(
 		(state: any) => state.axiosProcess.selectedFile
 	);
@@ -28,10 +30,6 @@ export default function PlainContext(props) {
 	} else {
 		classes = 'hidden';
 	}
-
-	const handleOnLeave = () => {
-		dispatch(setRenderConditionFalse());
-	};
 
 	const contextManager = (type: string) => {
 		switch (type) {
@@ -53,7 +51,7 @@ export default function PlainContext(props) {
 				top: props.top,
 				left: props.left,
 			}}
-			onMouseLeave={handleOnLeave}
+			onMouseLeave={logic.setRenderConditionFalse}
 		>
 			{contextManager(selectedFile.type)}
 		</div>
