+++ src/components/mapping/folderMapping/FolderListLogic.ts
@@ -3,6 +3,8 @@ import {
 	setPosition,
 	setRenderConditionTrue,
 } from '../../../redux/contextSlice';
+import axiosHelper from '@/logic/axiosHelper.js';
+import { setCurrentFolder, setPath } from '@/redux/currentSlice';
 
 export default class FolderListLogic {
 	constructor(store) {
@@ -31,4 +33,20 @@ export default class FolderListLogic {
 		this.store.dispatch(setPosition({ x: x, y: y }));
 		this.store.dispatch(setRenderConditionTrue());
 	};
+
+	openFolder(folderId) {
+		const currentPath = this.store.getState().current.currentPath;
+		const parsedCurrentPath = JSON.parse(JSON.stringify(currentPath));
+		parsedCurrentPath.push(folderId);
+		this.store.dispatch(setPath(parsedCurrentPath));
+		axiosHelper.getFolder(folderId).then((response) => {
+			const newCurrentFolder = {
+				id: response.data.data.id,
+				folder_name: response.data.data.folder_name,
+				nested_folders: response.data.data.nested_folders,
+				items: response.data.data.items,
+			};
+			this.store.dispatch(setCurrentFolder(newCurrentFolder));
+		});
+	}
 }
