+++ tsconfig.json
@@ -15,8 +15,13 @@
 		"isolatedModules": true,
 		"noEmit": true,
 		"jsx": "react-jsx",
-		"noImplicitAny": false
+		"noImplicitAny": false,
+		"baseUrl": "./",
+		"paths": {
+			"@/*": ["src/*"]
+		}
 	},
-	"include": ["src"],
+
+	//"include": ["src"],
 	"references": [{ "path": "./tsconfig.node.json" }]
 }
